chore(auth): move role and public decorator to auth module
