+++ fake-twitter-api/src/auth/decorators/index.ts
@@ -1,6 +1,7 @@
 /**
- * This file contains the decorators for the auth.
+ * This file contains the decorators for the auth module.
  */
 
 export * from './roles.decorator';
 export * from './current-user.decorator';
+export * from './public.decorator';
