--- fake-twitter-api/src/common/decorators/roles.decorator.ts
@@ -1,8 +1,0 @@
-/**
- * This file contains the roles decorator.
- */
-
-import { SetMetadata } from '@nestjs/common';
-
-export const ROLES_KEY = 'roles';
-export const Roles = (...roles: string[]) => SetMetadata(ROLES_KEY, roles);
