+++ fake-twitter-api/src/common/guards/global-auth.guard.ts
@@ -5,7 +5,7 @@
 
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { IS_PUBLIC_KEY } from '../decorators/public.decorator';
+import { IS_PUBLIC_KEY } from '../../auth/decorators/public.decorator';
 
 @Injectable()
 export class GlobalAuthGuard implements CanActivate {
